+++ docs/user_guide/revvity/index.md
@@ -1,0 +1,12 @@
+# Revvity
+
+```{toctree}
+:maxdepth: 1
+
+celigo/index
+```
+
+PyLabRobot supports the following Revvity instruments:
+
+- [Celigo](celigo/index.md), an image cytometer with brightfield and fluorescence
+  imaging, autofocus, auto-exposure, and multi-field scanning.
