+++ pylabrobot/revvity/__init__.py
@@ -1,0 +1,5 @@
+"""Revvity instruments."""
+
+from pylabrobot.revvity.celigo import Celigo, CeligoConfig
+
+__all__ = ["Celigo", "CeligoConfig"]
